// src/models/experience-register.d.ts
export interface IExperienceRegister {
  RowNumber: number;
  register_id: number;
  branch_id: number;
  fullname: string;
  gender: number;
  level: string;
  date: Date;
  phone_number: string;
  address: string;
  created_date_time: Date;
  detail: string;
  status: number;
  lu_user_id: string;
}